move fhir conversion logic into context
