+++ backend/src/main/java/gov/cdc/usds/simplereport/config/AzureTestEventReportingQueueConfiguration.java
@@ -15,9 +15,7 @@ import gov.cdc.usds.simplereport.properties.AzureStorageQueueReportingProperties
 import gov.cdc.usds.simplereport.service.AzureStorageQueueFhirReportingService;
 import gov.cdc.usds.simplereport.service.AzureStorageQueueTestEventReportingService;
 import gov.cdc.usds.simplereport.service.TestEventReportingService;
-
 import java.util.concurrent.CompletableFuture;
-
 import lombok.Builder;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -32,152 +30,155 @@ import org.springframework.context.annotation.Profile;
 @Slf4j
 @Configuration
 class AzureTestEventReportingQueueConfiguration {
-    @Bean("csvQueueReportingService")
-    @ConditionalOnProperty(
-            value = "simple-report.azure-reporting-queue.enabled",
-            havingValue = "true")
-    TestEventReportingService storageQueueReportingService(
-            ObjectMapper mapper, @Qualifier("csvQueue") QueueAsyncClient queueClient) {
-        log.info("Configured for queue={}", queueClient.getQueueName());
-        return new AzureStorageQueueTestEventReportingService(mapper, queueClient);
+  @Bean("csvQueueReportingService")
+  @ConditionalOnProperty(
+      value = "simple-report.azure-reporting-queue.enabled",
+      havingValue = "true")
+  TestEventReportingService storageQueueReportingService(
+      ObjectMapper mapper, @Qualifier("csvQueue") QueueAsyncClient queueClient) {
+    log.info("Configured for queue={}", queueClient.getQueueName());
+    return new AzureStorageQueueTestEventReportingService(mapper, queueClient);
+  }
+
+  @Bean("fhirQueueReportingService")
+  @ConditionalOnProperty(
+      value = "simple-report.azure-reporting-queue.fhir-queue-enabled",
+      havingValue = "true")
+  TestEventReportingService fhirQueueReportingService(
+      FhirContext context,
+      @Qualifier("fhirQueueClient") QueueAsyncClient queueClient,
+      GitProperties gitProperties,
+      BulkTestResultUploadFhirConverter bulkTestResultUploadFhirConverter) {
+    log.info("Configured for queue={}", queueClient.getQueueName());
+    return new AzureStorageQueueFhirReportingService(
+        context, queueClient, gitProperties, bulkTestResultUploadFhirConverter);
+  }
+
+  @Bean
+  FhirContext fhirContext() {
+    return FhirContext.forR4();
+  }
+
+  @Profile(PROD)
+  @Primary
+  @Bean(name = "csvQueueReportingService")
+  @ConditionalOnMissingBean(name = "csvQueueReportingService")
+  TestEventReportingService noOpCSVReportingService() {
+    return NoOpCovidReportingService.builder().build();
+  }
+
+  @Profile(PROD)
+  @Bean(name = "fhirQueueReportingService")
+  @ConditionalOnMissingBean(name = "fhirQueueReportingService")
+  TestEventReportingService noOpFhirReportingService(
+      FhirContext context,
+      GitProperties gitProperties,
+      BulkTestResultUploadFhirConverter bulkTestResultUploadFhirConverter) {
+    return NoOpFHIRReportingService.builder()
+        .fhirContext(context)
+        .gitProperties(gitProperties)
+        .bulkTestResultUploadFhirConverter(bulkTestResultUploadFhirConverter)
+        .build();
+  }
+
+  @Profile("!" + PROD)
+  @Primary
+  @Bean(name = "csvQueueReportingService")
+  @ConditionalOnMissingBean(name = "csvQueueReportingService")
+  TestEventReportingService noOpDebugCSVReportingService() {
+    return NoOpCovidReportingService.builder().printSerializedTestEvent(true).build();
+  }
+
+  @Profile("!" + PROD)
+  @Bean(name = "fhirQueueReportingService")
+  @ConditionalOnMissingBean(name = "fhirQueueReportingService")
+  TestEventReportingService noOpDebugFhirReportingService(
+      FhirContext context,
+      GitProperties gitProperties,
+      BulkTestResultUploadFhirConverter bulkTestResultUploadFhirConverter) {
+    return NoOpFHIRReportingService.builder()
+        .fhirContext(context)
+        .gitProperties(gitProperties)
+        .bulkTestResultUploadFhirConverter(bulkTestResultUploadFhirConverter)
+        .printSerializedTestEvent(true)
+        .build();
+  }
+
+  @Bean("csvQueue")
+  @ConditionalOnProperty(
+      value = "simple-report.azure-reporting-queue.enabled",
+      havingValue = "true")
+  QueueAsyncClient queueServiceAsyncClient(AzureStorageQueueReportingProperties properties) {
+    return new QueueClientBuilder()
+        .connectionString(properties.getConnectionString())
+        .queueName(properties.getName())
+        .buildAsyncClient();
+  }
+
+  @Bean("fhirQueueClient")
+  @ConditionalOnProperty(
+      value = "simple-report.azure-reporting-queue.fhir-queue-enabled",
+      havingValue = "true")
+  QueueAsyncClient fhirQueueServiceAsyncClient(AzureStorageQueueReportingProperties properties) {
+    return new QueueClientBuilder()
+        .connectionString(properties.getConnectionString())
+        .queueName(properties.getFhirQueueName())
+        .buildAsyncClient();
+  }
+
+  @Builder
+  static class NoOpCovidReportingService implements TestEventReportingService {
+
+    @Builder.Default private boolean printSerializedTestEvent = false;
+
+    @Override
+    public CompletableFuture<Void> reportAsync(TestEvent testEvent) {
+      log.warn(
+          "No Covid TestEventReportingService configured; defaulting to no-op reporting for TestEvent [{}]",
+          testEvent.getInternalId());
+
+      if (printSerializedTestEvent) {
+        log.info("TestEvent serializes as: {}", toBuffer(testEvent));
+      }
+      return CompletableFuture.completedFuture(null);
     }
 
-    @Bean("fhirQueueReportingService")
-    @ConditionalOnProperty(
-            value = "simple-report.azure-reporting-queue.fhir-queue-enabled",
-            havingValue = "true")
-    TestEventReportingService fhirQueueReportingService(
-            FhirContext context,
-            @Qualifier("fhirQueueClient") QueueAsyncClient queueClient,
-            GitProperties gitProperties,
-            BulkTestResultUploadFhirConverter bulkTestResultUploadFhirConverter) {
-        log.info("Configured for queue={}", queueClient.getQueueName());
-        return new AzureStorageQueueFhirReportingService(
-                context, queueClient, gitProperties, bulkTestResultUploadFhirConverter);
+    private String toBuffer(TestEvent testEvent) {
+      try {
+        ObjectMapper mapper = new ObjectMapper();
+        return mapper.writeValueAsString(new TestEventExport(testEvent));
+      } catch (JsonProcessingException e) {
+        throw new TestEventSerializationFailureException(testEvent.getInternalId(), e.getMessage());
+      }
+    }
+  }
+
+  @Builder
+  static class NoOpFHIRReportingService implements TestEventReportingService {
+
+    @Builder.Default private boolean printSerializedTestEvent = false;
+
+    private FhirContext fhirContext;
+    private GitProperties gitProperties;
+    private BulkTestResultUploadFhirConverter bulkTestResultUploadFhirConverter;
+
+    @Override
+    public CompletableFuture<Void> reportAsync(TestEvent testEvent) {
+      log.warn(
+          "No FHIR TestEventReportingService configured; defaulting to no-op reporting for TestEvent [{}]",
+          testEvent.getInternalId());
+
+      if (printSerializedTestEvent) {
+        log.info("TestEvent bundled as: {}", toBuffer(testEvent));
+      }
+      return CompletableFuture.completedFuture(null);
     }
 
-    @Bean
-    FhirContext fhirContext() {
-        return FhirContext.forR4();
+    private String toBuffer(TestEvent testEvent) {
+      return fhirContext
+          .newJsonParser()
+          .encodeResourceToString(
+              bulkTestResultUploadFhirConverter.createFhirBundle(testEvent, gitProperties, "P"));
     }
-
-    @Profile(PROD)
-    @Primary
-    @Bean(name = "csvQueueReportingService")
-    @ConditionalOnMissingBean(name = "csvQueueReportingService")
-    TestEventReportingService noOpCSVReportingService() {
-        return NoOpCovidReportingService.builder().build();
-    }
-
-    @Profile(PROD)
-    @Bean(name = "fhirQueueReportingService")
-    @ConditionalOnMissingBean(name = "fhirQueueReportingService")
-    TestEventReportingService noOpFhirReportingService(
-            FhirContext context, GitProperties gitProperties, BulkTestResultUploadFhirConverter bulkTestResultUploadFhirConverter) {
-        return NoOpFHIRReportingService.builder()
-                .fhirContext(context)
-                .gitProperties(gitProperties)
-                .bulkTestResultUploadFhirConverter(bulkTestResultUploadFhirConverter)
-                .build();
-    }
-
-    @Profile("!" + PROD)
-    @Primary
-    @Bean(name = "csvQueueReportingService")
-    @ConditionalOnMissingBean(name = "csvQueueReportingService")
-    TestEventReportingService noOpDebugCSVReportingService() {
-        return NoOpCovidReportingService.builder().printSerializedTestEvent(true).build();
-    }
-
-    @Profile("!" + PROD)
-    @Bean(name = "fhirQueueReportingService")
-    @ConditionalOnMissingBean(name = "fhirQueueReportingService")
-    TestEventReportingService noOpDebugFhirReportingService(
-            FhirContext context, GitProperties gitProperties, BulkTestResultUploadFhirConverter bulkTestResultUploadFhirConverter) {
-        return NoOpFHIRReportingService.builder()
-                .fhirContext(context)
-                .gitProperties(gitProperties)
-                .bulkTestResultUploadFhirConverter(bulkTestResultUploadFhirConverter)
-                .printSerializedTestEvent(true)
-                .build();
-    }
-
-    @Bean("csvQueue")
-    @ConditionalOnProperty(
-            value = "simple-report.azure-reporting-queue.enabled",
-            havingValue = "true")
-    QueueAsyncClient queueServiceAsyncClient(AzureStorageQueueReportingProperties properties) {
-        return new QueueClientBuilder()
-                .connectionString(properties.getConnectionString())
-                .queueName(properties.getName())
-                .buildAsyncClient();
-    }
-
-    @Bean("fhirQueueClient")
-    @ConditionalOnProperty(
-            value = "simple-report.azure-reporting-queue.fhir-queue-enabled",
-            havingValue = "true")
-    QueueAsyncClient fhirQueueServiceAsyncClient(AzureStorageQueueReportingProperties properties) {
-        return new QueueClientBuilder()
-                .connectionString(properties.getConnectionString())
-                .queueName(properties.getFhirQueueName())
-                .buildAsyncClient();
-    }
-
-    @Builder
-    static class NoOpCovidReportingService implements TestEventReportingService {
-
-        @Builder.Default
-        private boolean printSerializedTestEvent = false;
-
-        @Override
-        public CompletableFuture<Void> reportAsync(TestEvent testEvent) {
-            log.warn(
-                    "No Covid TestEventReportingService configured; defaulting to no-op reporting for TestEvent [{}]",
-                    testEvent.getInternalId());
-
-            if (printSerializedTestEvent) {
-                log.info("TestEvent serializes as: {}", toBuffer(testEvent));
-            }
-            return CompletableFuture.completedFuture(null);
-        }
-
-        private String toBuffer(TestEvent testEvent) {
-            try {
-                ObjectMapper mapper = new ObjectMapper();
-                return mapper.writeValueAsString(new TestEventExport(testEvent));
-            } catch (JsonProcessingException e) {
-                throw new TestEventSerializationFailureException(testEvent.getInternalId(), e.getMessage());
-            }
-        }
-    }
-
-    @Builder
-    static class NoOpFHIRReportingService implements TestEventReportingService {
-
-        @Builder.Default
-        private boolean printSerializedTestEvent = false;
-
-        private FhirContext fhirContext;
-        private GitProperties gitProperties;
-        private BulkTestResultUploadFhirConverter bulkTestResultUploadFhirConverter;
-
-        @Override
-        public CompletableFuture<Void> reportAsync(TestEvent testEvent) {
-            log.warn(
-                    "No FHIR TestEventReportingService configured; defaulting to no-op reporting for TestEvent [{}]",
-                    testEvent.getInternalId());
-
-            if (printSerializedTestEvent) {
-                log.info("TestEvent bundled as: {}", toBuffer(testEvent));
-            }
-            return CompletableFuture.completedFuture(null);
-        }
-
-        private String toBuffer(TestEvent testEvent) {
-            return fhirContext
-                    .newJsonParser()
-                    .encodeResourceToString(bulkTestResultUploadFhirConverter.createFhirBundle(testEvent, gitProperties, "P"));
-        }
-    }
+  }
 }
