+++ backend/src/main/java/gov/cdc/usds/simplereport/service/fhirConversion/FhirConversionContext.java
@@ -1,0 +1,70 @@
+package gov.cdc.usds.simplereport.service.fhirConversion;
+
+import static gov.cdc.usds.simplereport.validators.CsvValidatorUtils.getIteratorForCsv;
+import static gov.cdc.usds.simplereport.validators.CsvValidatorUtils.getNextRow;
+
+import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.parser.IParser;
+import com.fasterxml.jackson.databind.MappingIterator;
+import gov.cdc.usds.simplereport.api.model.errors.CsvProcessingException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class FhirConversionContext {
+  private static FhirConversionStrategy selectedConversionStrategy;
+
+  public FhirConversionContext(FhirConversionStrategy conversionStrategy) {
+    selectedConversionStrategy = conversionStrategy;
+  }
+
+  final FhirContext ctx = FhirContext.forR4();
+  final IParser parser = ctx.newJsonParser();
+
+  public List<String> convertToFhirBundles(InputStream csvStream, UUID orgId) {
+    var futureTestEvents = new ArrayList<CompletableFuture<String>>();
+    final MappingIterator<Map<String, String>> valueIterator = getIteratorForCsv(csvStream);
+    while (valueIterator.hasNext()) {
+      final Map<String, String> row;
+      try {
+        row = getNextRow(valueIterator);
+      } catch (CsvProcessingException ex) {
+        // anything that would land here should have been caught and handled by the file validator
+        log.error("Unable to parse csv.", ex);
+        continue;
+      }
+
+      var future =
+          CompletableFuture.supplyAsync(
+                  () -> selectedConversionStrategy.convertRowToFhirBundle(row, orgId))
+              .thenApply(parser::encodeResourceToString);
+      futureTestEvents.add(future);
+    }
+
+    List<String> bundles =
+        futureTestEvents.stream()
+            .map(
+                future -> {
+                  try {
+                    return future.get();
+                  } catch (InterruptedException | ExecutionException e) {
+                    log.error("Bulk upload failure to convert to fhir.", e);
+                    Thread.currentThread().interrupt();
+                    throw new CsvProcessingException("Unable to process file.");
+                  }
+                })
+            .toList();
+
+    return bundles;
+  }
+}
