+++ backend/src/main/java/gov/cdc/usds/simplereport/service/fhirConversion/FhirConversionStrategy.java
@@ -1,14 +1,9 @@
 package gov.cdc.usds.simplereport.service.fhirConversion;
 
-import gov.cdc.usds.simplereport.api.model.filerow.FileRow;
+import java.util.Map;
+import java.util.UUID;
 import org.hl7.fhir.r4.model.Bundle;
 
-import java.io.InputStream;
-import java.util.List;
-import java.util.UUID;
-
 public interface FhirConversionStrategy {
-    Bundle convertRowToFhirBundle(FileRow csvRow, UUID orgId);
-
-    List<String> convertToFhirBundles(InputStream csvStream, UUID orgId);
+  Bundle convertRowToFhirBundle(Map<String, String> csvRow, UUID orgId);
 }
